requirements.txt
Django>=5.0,<6.0
psycopg[binary]>=3.2,<4.0
django-environ>=0.11,<0.12
python-dotenv>=1.0,<2.0
python-dateutil>=2.9,<3.0